requirements.txt
django==4.2.2
python-telegram-bot==13.7
environs==9.5.0